# restaurantapi-main/openapi/swagger-latest-team12.json
{
  "swagger": "2.0",
  "info": {
    "version": "2025-04-15T08:32:20Z",
    "title": "tm3-restaurant_api-sam5"
  },
  "host": "3rh91vhdtj.execute-api.eu-west-3.amazonaws.com",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
    }
  },
  "paths": {
...
    "/bookings/client": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "orders a table by client (Sprint 1: User Story 6)",
        "description": "This endpoint books a table",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Reservation request",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "locationId": {
                  "type": "string",
                  "example": "672846d5c951184d705b65d7"
                },
                "tableNumber": {
                  "type": "string",
                  "example": "1"
                },
                "date": {
                  "type": "string",
                  "format": "date",
                  "example": "2021-12-31"
                },
                "guestsNumber": {
                  "type": "string",
                  "example": "4"
                },
                "timeFrom": {
                  "type": "string",
                  "example": "12:15"
                },
                "timeTo": {
                  "type": "string",
                  "example": "14:00"
                }
              },
              "required": [
                "locationId",
                "tableNumber",
                "date",
                "guestsNumber",
                "timeFrom",
                "timeTo"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "locationAddress": {
                  "type": "string"
                },
                "date": {
                  "type": "string"
                },
                "timeSlot": {
                  "type": "string"
                },
                "preOrder": {
                  "type": "string"
                },
                "guestsNumber": {
                  "type": "string"
                },
                "feedbackId": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "id": "res-12345",
                "status": "CONFIRMED",
                "locationAddress": "123 Main Street, New York, NY 10001",
                "date": "2021-12-31",
                "timeSlot": "12:15 - 14:00",
                "preOrder": "None",
                "guestsNumber": "4",
                "feedbackId": ""
              }
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication token is missing or invalid",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      }
    },
    "/bookings/tables": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "returns a list of available tables (Sprint 1: User Story 5)",
        "description": "This endpoint returns a list of tables available for booking",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "query",
            "description": "Location ID (e.g., 672846d5c951184d705b65d7)",
            "required": false,
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "description": "Date (e.g., 2024-08-01)",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "time",
            "in": "query",
            "description": "Time (e.g., 12:00)",
            "required": false,
            "type": "string"
          },
          {
            "name": "guests",
            "in": "query",
            "description": "Number of guests (e.g., 4)",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "locationId": {
                    "type": "string"
                  },
                  "locationAddress": {
                    "type": "string"
                  },
                  "tableNumber": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "string"
                  },
                  "availableSlots": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      }
    },
    "/bookings/waiter": {
      "post": {
        "tags": [
          "Waiters"
        ],
        "summary": "assigns a waiter to a booking",
        "description": "This endpoint assigns a waiter to a booking. It requires authentication with appropriate permissions.",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Booking details for waiter assignment",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "clientType": {
                  "type": "string",
                  "description": "Type of client",
                  "enum": [
                    "VISITOR",
                    "REGISTERED"
                  ],
                  "example": "VISITOR"
                },
                "customerName": {
                  "type": "string",
                  "description": "Name of the customer",
                  "example": "gaurav"
                },
                "date": {
                  "type": "string",
                  "format": "date",
                  "description": "Date of the booking (YYYY-MM-DD)",
                  "example": "2025-04-19"
                },
                "guestsNumber": {
                  "type": "string",
                  "description": "Number of guests",
                  "example": "4"
                },
                "locationId": {
                  "type": "string",
                  "description": "ID of the restaurant location",
                  "example": "location-001"
                },
                "tableNumber": {
                  "type": "string",
                  "description": "Table number for the booking",
                  "example": "table-101"
                },
                "timeFrom": {
                  "type": "string",
                  "description": "Start time of the booking (HH:MM)",
                  "example": "10:30"
                },
                "timeTo": {
                  "type": "string",
                  "description": "End time of the booking (HH:MM)",
                  "example": "12:00"
                }
              },
              "required": [
                "clientType",
                "customerName",
                "date",
                "guestsNumber",
                "locationId",
                "tableNumber",
                "timeFrom",
                "timeTo"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Waiter successfully assigned to booking",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Booking successfully created"
                },
                "bookingId": {
                  "type": "string",
                  "example": "booking-12345"
                },
                "customerName": {
                  "type": "string",
                  "example": "gaurav"
                },
                "date": {
                  "type": "string",
                  "format": "date",
                  "example": "2025-04-19"
                },
                "timeSlot": {
                  "type": "string",
                  "example": "10:30 - 12:00"
                },
                "tableNumber": {
                  "type": "string",
                  "example": "table-101"
                },
                "locationId": {
                  "type": "string",
                  "example": "location-001"
                },
                "guestsNumber": {
                  "type": "string",
                  "example": "4"
                }
              }
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid input data"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication token is missing or invalid",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission to create bookings",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Insufficient permissions to create bookings"
                }
              }
            }
          },
          "404": {
            "description": "Not found - Location or table not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Location or table not found"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Table already booked for this time slot",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Table already booked for this time slot"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      }
    },
    "/bookings/waiter/{Id}": {
      "put": {
        "tags": [
          "Waiters"
        ],
        "summary": "assigns a waiter to a booking",
        "description": "This endpoint assigns a waiter to a specific booking. It requires authentication with appropriate permissions.",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "ID of the booking to assign a waiter to",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Booking details for waiter assignment",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "format": "date",
                  "description": "Date of the booking (YYYY-MM-DD)",
                  "example": "2025-04-24"
                },
                "timeFrom": {
                  "type": "string",
                  "description": "Start time of the booking (HH:MM)",
                  "example": "10:30"
                },
                "timeTo": {
                  "type": "string",
                  "description": "End time of the booking (HH:MM)",
                  "example": "12:00"
                },
                "tableNumber": {
                  "type": "string",
                  "description": "Table number for the booking",
                  "example": "table104"
                }
              },
              "required": [
                "date",
                "timeFrom",
                "timeTo",
                "tableNumber"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Waiter successfully assigned to booking",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Waiter successfully assigned to booking"
                },
                "bookingId": {
                  "type": "string"
                },
                "date": {
                  "type": "string",
                  "format": "date"
                },
                "timeSlot": {
                  "type": "string",
                  "example": "10:30 - 12:00"
                },
                "tableNumber": {
                  "type": "string"
                }
              }
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid input data"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication token is missing or invalid",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Not found - Booking or table not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Booking or table not found"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Table already booked for this time slot",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Table already booked for this time slot"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      }
    },
...
    "/waiter/{id}": {
      "delete": {
        "tags": [
          "Waiters"
        ],
        "summary": "deletes a specific waiter",
        "description": "This endpoint removes a waiter from the system. This may be a soft delete that changes their status to TERMINATED rather than completely removing the record.",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the waiter to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Waiter successfully deleted",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Waiter successfully deleted"
              }
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication token is missing or invalid",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Unauthorized"
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission to delete waiters",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Access denied: Admin role required"
              }
            }
          },
          "404": {
            "description": "Waiter not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Waiter with ID 'waiter-123' not found"
              }
            }
          },
          "409": {
            "description": "Conflict - Waiter has active assignments",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Cannot delete waiter with active assignments"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      }
    }
  },
...
}